♻️(store): move stores in store directory
