+++ src/pages/Experiences.svelte
@@ -3,7 +3,7 @@
 
   import type { Question } from '../components/QuestionsForm/QuestionsForm.svelte'
 
-  import { swiper } from '../stores'
+  import { swiper } from '../store'
   import Page from '../components/Page.svelte'
   import Experience from '../components/Experience.svelte'
   import PostExperience from './PostExperience.svelte'
