+++ src/store/swiper.ts
